+++ student.rb
@@ -1,0 +1,11 @@
+require_relative 'Person'
+
+class Student < Person
+  def classroom
+    @classroom = classroom
+  end
+
+  def play_hooky
+    "¯\(ツ)/¯"
+  end
+end
